+++ Magic/src/main/java/com/elmakers/mine/bukkit/wand/Wand.java
@@ -70,7 +70,7 @@ import com.elmakers.mine.bukkit.effect.SoundEffect;
 import com.elmakers.mine.bukkit.effect.WandEffectContext;
 import com.elmakers.mine.bukkit.effect.builtin.EffectRing;
 import com.elmakers.mine.bukkit.heroes.HeroesManager;
-import com.elmakers.mine.bukkit.item.ArmorSlot;
+import com.elmakers.mine.bukkit.item.InventorySlot;
 import com.elmakers.mine.bukkit.magic.BaseMagicConfigurable;
 import com.elmakers.mine.bukkit.magic.Mage;
 import com.elmakers.mine.bukkit.magic.MageClass;
@@ -5931,7 +5931,7 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
         if (slots != null) {
             for (String slotKey : slots) {
                 try {
-                    ArmorSlot slot = ArmorSlot.valueOf(slotKey.toUpperCase());
+                    InventorySlot slot = InventorySlot.valueOf(slotKey.toUpperCase());
                     if (slot.getSlot() == slotNumber) {
                         return true;
                     }
@@ -5948,7 +5948,7 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
         List<Integer> slots = null;
         if (getBoolean("wearable")) {
             slots = new ArrayList<>();
-            for (ArmorSlot slot : ArmorSlot.values()) {
+            for (InventorySlot slot : InventorySlot.values()) {
                 if (slot.getSlot() >= 0) {
                     slots.add(slot.getSlot());
                 }
@@ -5960,7 +5960,7 @@ public class Wand extends WandProperties implements CostReducer, com.elmakers.mi
             slots = new ArrayList<>();
             for (String slotKey : slotList) {
                 try {
-                    ArmorSlot slot = ArmorSlot.valueOf(slotKey.toUpperCase());
+                    InventorySlot slot = InventorySlot.valueOf(slotKey.toUpperCase());
                     slots.add(slot.getSlot());
                 } catch (Exception ex) {
                     controller.getLogger().warning("Invalid wearable slot: " + slotKey);